refactor: 로컬스토리지 키 상수화 파일 분리
