+++ src/constants/index.ts
@@ -1,5 +1,3 @@
-const CART_LIST_LOCAL_STORAGE_KEY = 'cartList';
-
 const ESC_KEY = 'Escape';
 
 const DEFAULT_MIN_COUNT = 1;
@@ -14,7 +12,6 @@ const CART_LIST_SKELETON_ITEM_LENGTH = 5;
 const SHIPPING_FEE = 3000;
 
 export {
-  CART_LIST_LOCAL_STORAGE_KEY,
   ESC_KEY,
   DEFAULT_MIN_COUNT,
   DEFAULT_MAX_COUNT,
