+++ src/constants/localStorage.ts
@@ -1,0 +1,4 @@
+const CART_LIST_LOCAL_STORAGE_KEY = 'CART_LIST';
+const CURRENT_SERVER_LOCAL_STORAGE_KEY = 'API_BASE_URL';
+
+export { CART_LIST_LOCAL_STORAGE_KEY, CURRENT_SERVER_LOCAL_STORAGE_KEY };
